backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.3.0
anthropic==0.25.8
boto3==1.34.95
weasyprint==62.1
django-ratelimit==4.1.0
bcrypt==4.1.3
django[bcrypt]>=5.0
python-slugify==8.0.4
celery==5.3.6
redis==5.0.4
sentry-sdk==1.45.0
gunicorn==22.0.0
django-storages==1.14.3
# V2
python-docx==1.1.2
